rpmem: fix potential integer overflow

+++ src/tools/rpmemd/rpmemd_config.c
@@ -468,7 +468,7 @@ parse_cl_args(int argc, char *argv[], struct rpmemd_config *config,
 		default:
 			if (set_option((enum rpmemd_option)opt, optarg, config)
 					== 0) {
-				*cl_options |= (uint64_t)(1 << opt);
+				*cl_options |= (1ULL << opt);
 			} else {
 				print_usage(argv[0]);
 				exit(-1);
